group repository classes in Repository folder
