+++ docs/0-3-laravel.md
@@ -1,10 +1,14 @@
 # Integrating with frameworks - Laravel
+
 ## Replacing the default repository type
+
 If you're choosing to use Doctrine with Laravel, there are some common Doctrine integration packages, however only some of them allow configuration of the default repository type.
 
 ### atrauzzi/laravel-doctrine
-For this package, in its configuration settings, set the `defaultRepository` setting to `Happyr\DoctrineSpecification\EntitySpecificationRepository`
+
+For this package, in its configuration settings, set the `defaultRepository` setting to `Happyr\DoctrineSpecification\Repository\EntitySpecificationRepository`
 
 ###  mitchellvanw/laravel-doctrine
-For this package, in its configuration settings, set the `repository` setting to `Happyr\DoctrineSpecification\EntitySpecificationRepository`
+
+For this package, in its configuration settings, set the `repository` setting to `Happyr\DoctrineSpecification\Repository\EntitySpecificationRepository`
